requirements.txt
validators
requests
tldextract
gitpython